freeware java 8.md
Prizm Network Node Not working properly on Windows 10/11 or Latest MacOS?
No problem, download the Java 8 ( freeware ) version after uninstalling the older version.

For ease of searching, we added the most stable version to github. (instructions inside)
Java 8 v221 for all Operating Systems ( Windows, Linux, MacOs )
In the period of transition to WEB and paid versions of Java We are already ahead, with freeWare Java8 and Web Cold Wallet!